refactor: organize screens under presentation
