+++ lib/presentation/profile/profile_screen.dart
@@ -6,7 +6,7 @@ import 'package:swappy/core/responsive/responsive.dart';
 import 'package:swappy/infrastructure/auth/auth_service.dart';
 import 'package:swappy/infrastructure/di/locator.dart';
 import 'package:swappy/presentation/widgets/adaptive_scaffold.dart';
-import 'package:swappy/screens/auth_gate.dart';
+import 'package:swappy/presentation/auth/auth_gate.dart';
 
 class ProfileScreen extends StatelessWidget {
   const ProfileScreen({Key? key}) : super(key: key);
